[step1] feat: PersonBuilder를 Person 내부로 응집
